+++ odbc-api/tests/integration.rs
@@ -3528,7 +3528,7 @@ fn detect_truncated_output_in_bulk_fetch(profile: &Profile) {
     assert!(matches!(
         cursor.fetch_with_truncation_check(true),
         Err(Error::TooLargeValueForBuffer {
-            indicator: Indicator::Length(10),
+            indicator: Some(10),
             buffer_index: 0,
         })
     ))